requirements.txt
datasets
pyahocorasick
transformers
pandas
scikit-learn
torch
accelerate
spacy
jupyter
notebook
matplotlib
umap-learn
protobuf
sentencepiece
